ref(profiling): move effect and mock import
