+++ static/app/views/profiling/profileDetails/profileDetails.tsx
@@ -6,12 +6,18 @@ import {t} from 'sentry/locale';
 import trackAdvancedAnalyticsEvent from 'sentry/utils/analytics/trackAdvancedAnalyticsEvent';
 import {useCurrentProjectFromRouteParam} from 'sentry/utils/profiling/hooks/useCurrentProjectFromRouteParam';
 import useOrganization from 'sentry/utils/useOrganization';
+import {useParams} from 'sentry/utils/useParams';
+import {ProfileGroupProvider} from 'sentry/views/profiling/profileGroupProvider';
+import {useProfiles} from 'sentry/views/profiling/profilesProvider';
 
 import {ProfileDetailsTable} from './components/profileDetailsTable';
 
 function ProfileDetails() {
   const organization = useOrganization();
   const currentProject = useCurrentProjectFromRouteParam();
+  const profiles = useProfiles();
+  const params = useParams();
+
   useEffect(() => {
     trackAdvancedAnalyticsEvent('profiling_views.profile_summary', {
       organization,
@@ -31,7 +37,12 @@ function ProfileDetails() {
       >
         <Layout.Body>
           <Layout.Main fullWidth>
-            <ProfileDetailsTable />
+            <ProfileGroupProvider
+              input={profiles.type === 'resolved' ? profiles.data : null}
+              traceID={params.eventID}
+            >
+              <ProfileDetailsTable />
+            </ProfileGroupProvider>
           </Layout.Main>
         </Layout.Body>
       </SentryDocumentTitle>
